Make build.{branch}.properties Optional
This update makes build.{branch}.properties an optional file. This makes it easier to spin-up a new branch/org without needing to create a branch-specific build properties file but still preserves branch-specific username/password environment variables by dynamically building them. However, the branch can still specify a branch-specific properties file.

For now, this change requires that the branch name match the environment variables. Hyphens ("-") and dots (".") are not supported for this, but I'm sure this could be modified to regex that stuff out.